requirements.txt
asgiref==3.4.1
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
mysql==0.0.3
mysqlclient==2.1.0
Pillow==8.4.0
psycopg2==2.9.2
sqlparse==0.4.2
tzdata==2021.5
